Move handle status to a file that can be shared with users and suppliers when creating a handle
