+++ app/_types/handle-available-status.ts
@@ -1,0 +1,8 @@
+export const status = {
+  Undefined: undefined,
+  Pending: 'Pending',
+  Error: 'Error',
+  Available: true,
+  Taken: false,
+} as const
+export type HandleStatus = (typeof status)[keyof typeof status]
